implement Blowfish - remove libmcrypt dependency
mcrypt doesn't seem to have a well maintained upstream and is already
deprecated by at least one package maintainer, see GitHub pull request #24.
Since it is not too hard to implement Blowfish ourselves,
we might as well remove the dependecy.

With gcc -O3, performance seems to be on par with libmcrypt.